Fix .htaccess when used with MultiViews
mod_rewrite + MultiViews don't play nice when the URL you want to redirect
is the same as a filename in the current folder (which applies for the
'list' and 'search' URLs).

The net result is that the first page of each works, but subsequent pagination
fails.

Disabling MultiViews solves this problem.